requirements.txt
Django==3.2.7
marshmallow==3.13.0

django-debug-toolbar==3.2.2
pytest-django==4.4.0
pytest-cov==2.12.1
mixer==7.1.2
pycountry==20.7.3